==> notebooks/3_Split_datasets.ipynb <==
# Split the dataset

import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split

#### Load dataset via Pickle

feature_vec = pd.read_pickle('feature_vec_opt_1m.pickle')

#### Reorder feature_vec to be <qid, label, features>

feature_vec = feature_vec[['Synthetic_match', 'Qid', 'Hist_Suffix_freq', 'ngram_1', 'ngram_2', 'ngram_3', 'ngram_4', 'ngram_5', 'ngram_6',
 'has_ne', 'ne_norm', 'candid_freq', 'prefixlen_char', 'suffixlen_char', 'totallen_char', 'prefixlen_word', 'suffixlen_word', 'totallen_word',
 'bool_space']]

feature_vec.columns = ['label', 'qid', 'f1', 'f2', 'f3', 'f4', 'f5', 'f6', 'f7', 'f8', 'f9', 'f10', 'f11',  'f12', 'f13', 'f14', 'f15', 'f16', 'f17']
feature_vec

train_df, test_df = train_test_split(feature_vec, test_size=0.2)

feature_vec['qid'] = feature_vec['qid'].astype('int')
feature_vec['label'] = feature_vec['label'].astype('int')

feature_vec = feature_vec.sort_values('qid')

feature_vec.to_pickle('feature_df.pickle')
